requirements.txt
streamlit
langchain
openai
fitz (PyMuPDF)
faiss-cpu
requests
PyMuPDF
langchain_community
python-dotenv
tiktoken
faiss-cpu
simple-salesforce
